% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_interface.R
\name{itemType}
\alias{itemType}
\title{List available flags under a given type}
\usage{
itemType(type)
}
\arguments{
\item{type}{A character string specifying the flag type.}
}
\value{
A character vector of available flags under the given type.
}
\description{
List available flags under a given type
}
\examples{
itemType("nation")
}
