% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_interface.R
\name{typeFlag}
\alias{typeFlag}
\title{List available flag types}
\usage{
typeFlag()
}
\value{
A character vector of available flag types.
}
\description{
This function returns all supported flag categories.
}
\examples{
typeFlag()
}
