% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_interface.R
\name{plotCNFlag}
\alias{plotCNFlag}
\title{Plot a Chinese national or party flag}
\usage{
plotCNFlag(type, item, ...)
}
\arguments{
\item{type}{Flag category. Available options can be queried via typeFlag().}

\item{item}{Specific flag name under the given type.}

\item{...}{Additional parameters passed to the underlying plotting function.}
}
\value{
A ggplot object.
}
\description{
This is the unified interface for plotting flags in ggChinaFlag.
}
\examples{
\dontrun{
plotCNFlag(type = "nation", item = "PRC")
}
}
